+++ rabbithole-derive/src/lib.rs
@@ -47,39 +47,60 @@ fn inner_derive(input: TokenStream) -> syn::Result<proc_macro2::TokenStream> {
                         self.#to_ones.to_resource_identifier(),
                         self.to_relationship_links(stringify!(#to_ones), uri)?,
                     ) {
-                        let data = IdentifierData::Single(Some(relat_id));
-                        let relat = Relationship { data, links, ..Default::default() };
+                        let data = rabbithole::model::resource::IdentifierData::Single(Some(relat_id));
+                        let relat = rabbithole::model::relationship::Relationship { data, links, ..std::default::Default::default() };
                         relat_map.insert(stringify!(#to_ones).to_string(), relat);
                     }
                 )*
 
                 #(
                     if let Some(links) = self.to_relationship_links(stringify!(#to_manys), uri)? {
-                        let mut relat_ids: ResourceIdentifiers = Default::default();
+                        let mut relat_ids: rabbithole::model::resource::ResourceIdentifiers = std::default::Default::default();
                         for item in &self.#to_manys {
                             if let Some(relat_id) = item.to_resource_identifier() {
                                 relat_ids.push(relat_id);
                             }
                         }
-                        let data = IdentifierData::Multiple(relat_ids);
-                        let relat = Relationship { data, links, meta: Default::default() };
+                        let data = rabbithole::model::resource::IdentifierData::Multiple(relat_ids);
+                        let relat = rabbithole::model::relationship::Relationship { data, links, meta: std::default::Default::default() };
                         relat_map.insert(stringify!(#to_manys).to_string(), relat);
                     }
                 )*
 
                 Ok(Some(relat_map))
             }
-            fn included(&self, uri: &str) -> rabbithole::RbhOptionRes<rabbithole::model::document::Included> {
+            fn included(&self, uri: &str,
+                include_query: &std::option::Option<rabbithole::model::query::IncludeQuery>,
+                fields_query: &std::option::Option<rabbithole::model::query::FieldsQuery>,
+            ) -> rabbithole::RbhOptionRes<rabbithole::model::document::Included> {
                 let mut included: rabbithole::model::document::Included = Default::default();
                 #(
-                    if let Some(res) = self.#to_ones.to_resource(uri)? {
-                        included.push(res);
+                    if let Some(included_fields) = include_query {
+                        if included_fields.contains(stringify!(#to_ones)) {
+                            if let Some(res) = self.#to_ones.to_resource(uri, fields_query)? {
+                                included.insert(res);
+                            }
+                        }
+                    } else {
+                        if let Some(res) = self.#to_ones.to_resource(uri, fields_query)? {
+                            included.insert(res);
+                        }
                     }
                 )*
                 #(
-                    for item in &self.#to_manys {
-                        if let Some(data) = item.to_resource(uri)? {
-                            included.push(data);
+                    if let Some(included_fields) = include_query {
+                        if included_fields.contains(stringify!(#to_manys)) {
+                            for item in &self.#to_manys {
+                                if let Some(data) = item.to_resource(uri, fields_query)? {
+                                    included.insert(data);
+                                }
+                            }
+                        }
+                    } else {
+                        for item in &self.#to_manys {
+                            if let Some(data) = item.to_resource(uri, fields_query)? {
+                                included.insert(data);
+                            }
                         }
                     }
                 )*