+++ rabbithole/src/model/query.rs
@@ -1,0 +1,37 @@
+use std::collections::{HashMap, HashSet};
+
+pub type IncludeQuery = HashSet<String>;
+pub type FieldsQuery = HashMap<String, HashSet<String>>;
+pub type SortQuery = Vec<String>;
+
+#[derive(Debug, Default)]
+pub struct Query {
+    /// When include is:
+    ///   1. `None`: all included fields will be added
+    ///   2. `Some(<empty-query>)`: no included field will be added
+    ///   3. `Some(<some-query>)`:
+    ///     1. No query matches the existing fields: like branch 2
+    ///     2. Otherwise: all matched fields will be added
+    pub include: Option<IncludeQuery>,
+    /// Just like `include`
+    pub fields: Option<FieldsQuery>,
+    /// When sort is:
+    ///   1. empty: no sorting at all, clients should not expect the order of the result
+    ///   2. some values, but none of the values matches: like branch 1
+    ///   3. some values, and some of the items matches: sorting result with the order of the matched sort-query item
+    pub sort: SortQuery,
+    pub page: Option<PageQuery>,
+    pub filter: Option<FilterQuery>,
+}
+
+#[derive(Debug)]
+pub enum PageQuery {
+    OffsetBased { offset: u32, limit: u32 },
+    PageBased { number: u32, size: u32 },
+    CursorBased(String),
+}
+
+#[derive(Debug)]
+pub enum FilterQuery {
+    Rsql(rsql_rs::ast::expr::Expr),
+}